Makefile: install click when creating venv
Not the cleanest solution but good enough for our development. sby
requires click, and installing click inside the venv inside the CI
would require a number of commands (create venv, activate it and
install click), instead of just calling "make concept_name". We want
our Makefile to be able to do things so we can focus on development, so
this solution looks good