CI: use uv run --dev when we need to have the coverage tool installed
uv is *very* powerful and it automatically and really fast creates the
specific environment that is needed. This means that it will *remove*
all development dependencies if it thinks they are not needed to run a
specific command!

Probably, in the near future we will have to stop manually managing the
venv because uv does it automatically and much better that what we can
manage with the Makefile

But for now, we need the CI to work :)