doc: downgrade readthedocs theme
Versions from 3.0 onward do not show the version number in the
generated doc unless you host it on readthedocs.

The issue is documented here:
https://github.com/readthedocs/sphinx_rtd_theme/issues/1624 .  In the
issue, there is a pull request that could solve it but it was submitted in Oct
2024 and we are on Nov 2025 and they haven't reviewed/accepted it yet,
so it seems unlikely they'll accept it anytime soon.

Thus the only option was to downgrade the sphinx theme. In the future
we may consider changing it for a different theme